\name{Juan Diego Cubides}
\alias{hello}
\title{Primera Entrega Area Electronica!}
\usage{
library(nycflights13)
library(tidyverse)
library(dplyr)
}
\description{
Somos estudiantes de Ingenieria Mecatronica de la universidad ECCI,
Este repositorio contiene ejerecicios para practicar las funciones de transformacion en rstudio y un pdf realizado en R makdown
}
\examples{

}
